Finalize refactor of response extraction
