+++ app/app_test.go
@@ -5,7 +5,7 @@ import (
 	"encoding/json"
 	"github.com/mortedecai/gbb/gbb"
 	"github.com/mortedecai/gbb/gbb/mocks"
-	response2 "github.com/mortedecai/gbb/gbb/response"
+	"github.com/mortedecai/gbb/response"
 	"go.uber.org/mock/gomock"
 	"io"
 	"net/http"
@@ -123,10 +123,10 @@ var _ = Describe("App", func() {
 })
 
 func createEmptyDownloadResponse() *http.Response {
-	basicResponse := response2.GBBDownloadFilesResponse{
+	basicResponse := response.GBBDownloadFilesResponse{
 		Success: true,
-		Data: response2.GBBFileDownloadData{
-			Files: make([]response2.GBBDownloadFile, 0),
+		Data: response.GBBFileDownloadData{
+			Files: make([]response.GBBDownloadFile, 0),
 		},
 	}
 	data, err := json.Marshal(basicResponse)
