+++ gbb/gbb.go
@@ -4,8 +4,8 @@ import (
 	"bytes"
 	"encoding/json"
 	"fmt"
-	"github.com/mortedecai/gbb/gbb/response"
 	"github.com/mortedecai/gbb/gbberror"
+	"github.com/mortedecai/gbb/response"
 	"io"
 	"net/http"
 	"os"
